// src/app/core/services/model/thermalDisinfectionTaskStatusEnum.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ThermalDisinfectionTaskStatusEnum = 'Idle' | 'CheckingBlockedStatus' | 'CheckingForActivatedKeySwitch' | 'SettingCirculationValves' | 'WaitingForDteTemperatureClaimByUser' | 'SettingDteTemperatureFirstTry' | 'SettingDteTemperatureSecondTry' | 'WaitingForKeySwitchToTurnOn' | 'WaitingForKeySwitchToTurnOffTurnOn' | 'FlushingDevices' | 'Successful' | 'Failed';

export const ThermalDisinfectionTaskStatusEnum = {
    Idle: 'Idle' as ThermalDisinfectionTaskStatusEnum,
    CheckingBlockedStatus: 'CheckingBlockedStatus' as ThermalDisinfectionTaskStatusEnum,
    CheckingForActivatedKeySwitch: 'CheckingForActivatedKeySwitch' as ThermalDisinfectionTaskStatusEnum,
    SettingCirculationValves: 'SettingCirculationValves' as ThermalDisinfectionTaskStatusEnum,
    WaitingForDteTemperatureClaimByUser: 'WaitingForDteTemperatureClaimByUser' as ThermalDisinfectionTaskStatusEnum,
    SettingDteTemperatureFirstTry: 'SettingDteTemperatureFirstTry' as ThermalDisinfectionTaskStatusEnum,
    SettingDteTemperatureSecondTry: 'SettingDteTemperatureSecondTry' as ThermalDisinfectionTaskStatusEnum,
    WaitingForKeySwitchToTurnOn: 'WaitingForKeySwitchToTurnOn' as ThermalDisinfectionTaskStatusEnum,
    WaitingForKeySwitchToTurnOffTurnOn: 'WaitingForKeySwitchToTurnOffTurnOn' as ThermalDisinfectionTaskStatusEnum,
    FlushingDevices: 'FlushingDevices' as ThermalDisinfectionTaskStatusEnum,
    Successful: 'Successful' as ThermalDisinfectionTaskStatusEnum,
    Failed: 'Failed' as ThermalDisinfectionTaskStatusEnum
};